validate-config/requirements.txt
pathways-typing @ git+https://github.com/BLSQ/pathways-typing@main